Move `Gc<T>` smart pointer API to own its own file
We make sure to re-export `Gc` in lib.rs so that it's still available
externally via gcmalloc::Gc;
